+++ changelog.md
@@ -4,6 +4,7 @@
 <!-- Add all new changes here. They will be moved under a version at release -->
 * `FIX` cannot debug in Linux due to lua-debug expecting host process to have lua54 symbols available
 * `FIX` support hex color codes with `#` in `textDocument/documentColor`
+* `NEW` type reference from `getmetatable` is supported now
 
 ## 3.14.0
 `2025-4-7`